requirements.txt
MySQL-python>=1.2.4
PIL>=1.1.7
Django>=1.5.1

#
# used by Django for timezone support
#
pytz

South>=0.7.6
PyYAML>=3.10
django-crispy-forms>=1.2.8
boto>=2.9.2
django-debug-toolbar>=0.9.4
nose>=1.3.0
django-nose>=1.1
coverage>=3.6
sqlparse>=0.1.7
mmh3>=2.2
requests>=1.2.0
python-nmap>=0.2.7
django-tastypie>=0.9.15
django-celery>=3.0.17
texttable>=0.8.1
-e hg+https://bitbucket.org/catherinedevlin/cmd2#egg=cmd2
django-bootstrap-pagination>=0.1.10
factory-boy>=2.0.2
django-model-utils>=1.4.0

